resolved compile errors on testcase around LinearProbeHashXXXX. but TestHashTable fails...

+++ container/hash/hash_table_page_test.go
@@ -46,15 +46,15 @@ func TestHashTableHeaderPage(t *testing.T) {
 	// add a few hypothetical block pages
 	for i := 0; i < 10; i++ {
 		headerPage.AddBlockPageId(types.PageID(i))
-		if i+1 != headerPage.NumBlocks() {
+		if uint32(i+1) != headerPage.NumBlocks() {
 			t.Errorf("NumBlocks shoud be %d, but got %d", i+1, headerPage.NumBlocks())
 		}
 	}
 
 	// check for correct block page IDs
 	for i := 0; i < 10; i++ {
-		if types.PageID(i) != headerPage.GetBlockPageId(i) {
-			t.Errorf("GetBlockPageId shoud be %d, but got %d", i, headerPage.GetBlockPageId(i))
+		if types.PageID(i) != headerPage.GetBlockPageId(uint32(i)) {
+			t.Errorf("GetBlockPageId shoud be %d, but got %d", i, headerPage.GetBlockPageId(uint32(i)))
 		}
 	}
 
@@ -75,32 +75,32 @@ func TestHashTableBlockPage(t *testing.T) {
 	blockPage := (*page.HashTableBlockPage)(unsafe.Pointer(newPageData))
 
 	for i := 0; i < 10; i++ {
-		blockPage.Insert(i, i, i)
+		blockPage.Insert(uint32(i), uint32(i), uint32(i))
 	}
 
 	for i := 0; i < 10; i++ {
 		// equals(t, i, blockPage.KeyAt(i))
-		testingpkg.Assert(t, i == blockPage.KeyAt(i), "")
+		testingpkg.Assert(t, uint32(i) == blockPage.KeyAt(uint32(i)), "")
 		//equals(t, i, blockPage.ValueAt(i))
-		testingpkg.Assert(t, i == blockPage.ValueAt(i), "")
+		testingpkg.Assert(t, uint32(i) == blockPage.ValueAt(uint32(i)), "")
 	}
 
 	for i := 0; i < 10; i++ {
 		if i%2 == 1 {
-			blockPage.Remove(i)
+			blockPage.Remove(uint32(i))
 		}
 	}
 
 	for i := 0; i < 15; i++ {
 		if i < 10 {
-			testingpkg.Assert(t, true == blockPage.IsOccupied(i), "block page should be occupied")
+			testingpkg.Assert(t, true == blockPage.IsOccupied(uint32(i)), "block page should be occupied")
 			if i%2 == 1 {
-				testingpkg.Assert(t, false == blockPage.IsReadable(i), "block page should not be readable")
+				testingpkg.Assert(t, false == blockPage.IsReadable(uint32(i)), "block page should not be readable")
 			} else {
-				testingpkg.Assert(t, true == blockPage.IsReadable(i), "block page should be readable")
+				testingpkg.Assert(t, true == blockPage.IsReadable(uint32(i)), "block page should be readable")
 			}
 		} else {
-			testingpkg.Assert(t, false == blockPage.IsOccupied(i), "block page should not be occupied")
+			testingpkg.Assert(t, false == blockPage.IsOccupied(uint32(i)), "block page should not be occupied")
 		}
 	}
